+++ src/Domain/Repositories/IHoldingRepository.cs
@@ -10,7 +10,7 @@ public interface IHoldingRepository
         Guid stockId,
         CancellationToken cancellationToken = default);
     Task<IReadOnlyList<Holding>> GetByPortfolioIdAsync(Guid portfolioId, CancellationToken cancellationToken = default);
-    Task<int> GetTotalQuantityByStockAsync(Guid stockId, CancellationToken cancellationToken = default);
+    Task<decimal> GetTotalQuantityByStockAsync(Guid stockId, CancellationToken cancellationToken = default);
     Task AddAsync(Holding holding, CancellationToken cancellationToken = default);
     void Update(Holding holding);
     void Remove(Holding holding);
